// apps/web/src/meloming/domains/channel/types/song.ts
export type PostSongsChannelIdentifierBulkRequestBody = {
  songs: Array<
    {
      title: string;
      albumArt?: string;
      karaokeUrl?: string;
      coverUrl?: string | null;
      originalUrl?: string | null;
      difficulty?: number;
      proficiency?: number;
      songKey?: string;
      bpm?: number | null;
      lyricsLink?: string | null;
      lyricsText?: string | null;
      description?: string | null;
      currencyPrices?: Record<string, number | null> | null;
      autoSearchAlbumArt?: boolean;
    } & (
      | {
          artistName: string;
          categoryNames?: string[];
        }
      | {
          artistId: number;
          categoryIds?: number[];
        }
    )
  >;
};